feat: static folder for js / css & assets (img) implemented
A static folder was created within the main app with the subfolders css, js & img, where all css / js & assets are stored. The static folder was integrated in the settings.py with its path and directory. In the base and login.htm, the corresponding style.css was integrated in the head area using {% load static %}.