feat: register.html & corresponding view implemented
An HTML template "register" was created. The end user has the option to register there. To do this, he must enter his username, email and desired password and confirm them again to rule out the possibility of a typo. There is also a database query to see if the entered username/email is already in use. In this case, the user receives feedback that the entered username/email is already in use. If the password and the confirmed password do not match, this is also displayed to the user.