refactor: index html, method of form adjusted and JS function to post messages implemented
Within the index.html, the "action" method in the form was replaced by the onsubmit method, which calls a new JS function.

The new asynchronous JS function "sendMessage" was implemented. Within the function, the variable "formdata - fd" is defined and is assigned the value of our text input field. A variable token was also defined; the value corresponds to the django csrfmiddlewaretoken.

Within the function there is a try-catch block. The path /chat/ is caught as annchronously and the formdata in the body is passed as a payload using a POST request. There is then a console output that the addition of the new message was successful. If an error occurs, the catch block is addressed and is passed the error as a parameter, which is then logged out as a concole.warn statement.