refactor: index.html & sendMessage() adjusted with DOM manipulation to render data from frontend while / after sending a new message
The JavaScript function sendMessage() has been adapted to display the message sent from the frontend in the chat history using JS. To do this, the DOM is manipulated. Before communication with the backend, the message is displayed in gray using innerHTML += the history. As soon as communication with the backend has been successful, the generated HTML element is deleted and then recreated, this time with black text to let the user know that their message was sent successfully. Currently, only the newly added data from the frontend is rendered first; in the next step, the data from the backend is accessed.