refactor: views.py & index.html adjusted to send JsonResponse
def index in the views.py was adjusted so that the response that is returned is a JsonResponse. the variable new_message receives the value of the new message object. the object is then serialized using serializers. the JsonResponse is then returned where we return a substring and remove the first and last element of the array. in the index-html sendMessage() the variable resonse was assigned the value of our server response. the variable json is assigned the value of the response in an asynchronous operation in connection with the function .json(). the variable parsedJoson is then assigned the value of the variable json and the function JSON.pase(json) is used to convert the text json into a real json object