refactor: register.html
the register.html has been adapted, instead of working directly with "action=backend_url" when sending the data, a JavaSript function has now been implemented onsubmit "registerNewUser()". The function initially deactivates all input fields and shows a loading animation while waiting for the response from the backend. If the registration was successful, the user is forwarded to the /chat route.